DBase/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.8.2
kafka-python==2.0.2
redis==3.5.3
pymongo==3.12.0
psycopg2-binary==2.9.1
python-dotenv==0.19.0
requests==2.26.0